push newest

+++ src/main/java/com/longpt37/MuleWebAPILongPT37/embeded/PurchaseOrderDetailsId.java
@@ -8,5 +8,5 @@ import java.io.Serializable;
 public class PurchaseOrderDetailsId implements Serializable {
     private String orderNo;
     private String productCode;
-    private float purchasePrice;
+    private double purchasePrice;
 }
