+++ src/main/java/com/longpt37/MuleWebAPILongPT37/embeded/SalesOrderDetailsId.java
@@ -8,5 +8,5 @@ import java.io.Serializable;
 public class SalesOrderDetailsId implements Serializable {
     private String orderNo;
     private String productCode;
-    private float salesPrice;
+    private double salesPrice;
 }
